move headers to its own folder inside components
